rename some files to lowercase
